# internal/models/portfolio.go
package models

import (
	"time"
)

// Portfolio represents a collection of addresses grouped together
type Portfolio struct {
	ID          string    `json:"id" db:"id"`
	UserID      string    `json:"userId" db:"user_id"`
	Name        string    `json:"name" db:"name"`
	Description *string   `json:"description,omitempty" db:"description"`
	Addresses   []string  `json:"addresses" db:"addresses"`
	CreatedAt   time.Time `json:"createdAt" db:"created_at"`
	UpdatedAt   time.Time `json:"updatedAt" db:"updated_at"`
}